initial _BaseROM split
